requirements.txt
google-generativeai
pyperclip
keyboard
PyQt6
python-dotenv
requests
